requirements.txt
argcomplete==1.12.3
attrs==21.2.0
awscli==1.20.65
boto3==1.18.41
botocore==1.21.65
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
cycler==0.11.0
dill==0.3.4
docutils==0.15.2
durationpy==0.5
Flask==2.0.2
fonttools==4.28.4
future==0.18.2
google-pasta==0.2.0
hjson==3.0.2
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kappa==0.6.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
multiprocess==0.70.12.2
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pathos==0.2.8
pep517==0.12.0
Pillow==8.4.0
pip-tools==6.4.0
placebo==0.10.0
pox==0.3.0
ppft==1.6.6.4
protobuf==3.19.1
protobuf3-to-dict==0.1.5
pyasn1==0.4.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
sagemaker==2.65.0
scikit-image==0.18.1
scipy==1.7.3
six==1.16.0
smdebug-rulesconfig==1.0.1
text-unidecode==1.3
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.0
tqdm==4.62.3
troposphere==3.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wsgi-request-logger==0.4.6
zappa==0.54.1
zipp==3.6.0
